pyproject.toml
[tool.poetry]
name = "taf_tasks"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
yoyo-migrations = "^7.3.2"
aiopg = "^1.3.4"
dependency-injector = "^4.40.0"
requests = "^2.28.1"
uvicorn = "^0.18.3"
fastapi = "^0.82.0"
aiohttp = "^3.8.1"
DateTime = "^4.7"
numpy = "^1.23.5"
httpx = "^0.23.1"
websockets = "^10.4"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
isort = "^5.10.1"
black = "^22.1.0"
docformatter = "^1.4"
add-trailing-comma = "^2.2.1"
flake8 = "^5.0.4"
mypy = "^0.971"
safety = "^2.1.1"
bandit = "^1.7.1"
itsdangerous = "^2.0.1"
MarkupSafe = "^2.1.0"
Sphinx = "^4.4.0"
sphinx-pydantic = "^0.1.1"
sphinx-rtd-theme = "^1.0.0"
pytest-asyncio = "^0.19.0"
pytest = "^7.0.1"
#coverage = "^6.4.4"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
